ci: remove prefix mpv-
