+++ mpv-autoload/meta/autoload.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-autoload
+// @name         autoload
 // @version      0.1.13
 // @description  mpv player autoload script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-autoload.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/autoload.js
 // ==/UserScript==
