+++ mpv-autoload/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-autoload.meta.js"
-const files = ["./bundle/mpv-autoload.js", "./es5/mpv-autoload.js"]
+const meta = "./meta/autoload.meta.js"
+const files = ["./bundle/autoload.js", "./es5/autoload.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
