+++ mpv-clipboard-play/build.ts
@@ -2,7 +2,7 @@ import { build } from "esbuild"
 import fs from "node:fs"
 
 build({
-  entryPoints: ["./src/mpv-clipboard-play.ts"],
+  entryPoints: ["./src/clipboard-play.ts"],
   bundle: true,
   outdir: "bundle",
   minify: true,
