+++ mpv-clipboard-play/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-clipboard-play.meta.js"
-const files = ["./bundle/mpv-clipboard-play.js", "./es5/mpv-clipboard-play.js"]
+const meta = "./meta/clipboard-play.meta.js"
+const files = ["./bundle/clipboard-play.js", "./es5/clipboard-play.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
