+++ mpv-clipboard-play/meta/clipboard-play.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-clipboard-play
+// @name         clipboard-play
 // @version      0.1.13
 // @description  mpv player clipboard-play script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-clipboard-play.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/clipboard-play.js
 // ==/UserScript==
