+++ mpv-copy-screen/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-copy-screen.meta.js"
-const files = ["./bundle/mpv-copy-screen.js", "./es5/mpv-copy-screen.js"]
+const meta = "./meta/copy-screen.meta.js"
+const files = ["./bundle/copy-screen.js", "./es5/copy-screen.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
