+++ mpv-copy-screen/meta/copy-screen.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-copy-screen
+// @name         copy-screen
 // @version      0.1.13
 // @description  mpv player copy-screen script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-copy-screen.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/copy-screen.js
 // ==/UserScript==
