+++ mpv-copy-screen/readme.md
@@ -2,7 +2,7 @@
 Use the shortcut (ctrl+c) to copy the current video content to the clipboard
 
 ## install
-- Download the latest version of [mpv-copy-screen.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory
+- Download the latest version of [copy-screen.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory
 - Add configuration to input.conf [shortkey](https://github.com/mpv-easy/mpv-easy/tree/main/mpv-copy-screen#shortkey)
 
 ## shortkey
