+++ mpv-copy-time/meta/copy-time.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-copy-time
+// @name         copy-time
 // @version      0.1.13
 // @description  mpv player copy-time script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-copy-time.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/copy-time.js
 // ==/UserScript==
