+++ mpv-copy-time/readme.md
@@ -2,7 +2,7 @@
 Use the shortcut (ctrl+C) to copy the current video time to the clipboard
 
 ## install
-- Download the latest version of [mpv-copy-time.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory
+- Download the latest version of [copy-time.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory
 - Add configuration to input.conf [shortkey](https://github.com/mpv-easy/mpv-easy/tree/main/mpv-copy-time#shortkey)
 -
