+++ mpv-crop/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-crop.meta.js"
-const files = ["./bundle/mpv-crop.js", "./es5/mpv-crop.js"]
+const meta = "./meta/crop.meta.js"
+const files = ["./bundle/crop.js", "./es5/crop.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
