+++ mpv-crop/meta/crop.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-meta
+// @name         crop
 // @version      0.1.13
 // @description  mpv player crop script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-crop.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/crop.js
 // ==/UserScript==
