+++ mpv-cut/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-cut.meta.js"
-const files = ["./bundle/mpv-cut.js", "./es5/mpv-cut.js"]
+const meta = "./meta/cut.meta.js"
+const files = ["./bundle/cut.js", "./es5/cut.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
