+++ mpv-cut/meta/cut.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-cut
+// @name         cut
 // @version      0.1.13
 // @description  mpv player cut script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-cut.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/cut.js
 // ==/UserScript==
