+++ mpv-translate/build.ts
@@ -2,7 +2,7 @@ import { build } from "esbuild"
 import fs from "node:fs"
 
 build({
-  entryPoints: ["./src/mpv-translate.tsx"],
+  entryPoints: ["./src/translate.tsx"],
   bundle: true,
   outdir: "bundle",
   minify: true,
