+++ mpv-translate/meta/build.ts
@@ -1,7 +1,7 @@
 import { writeFileSync, readFileSync } from "node:fs"
 
-const meta = "./meta/mpv-translate.meta.js"
-const files = ["./bundle/mpv-translate.js", "./es5/mpv-translate.js"]
+const meta = "./meta/translate.meta.js"
+const files = ["./bundle/translate.js", "./es5/translate.js"]
 
 const metaStr = readFileSync(meta)
 for (const p of files) {
