+++ mpv-translate/meta/translate.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
-// @name         mpv-translate
+// @name         translate
 // @version      0.1.13
 // @description  mpv player translation script
 // @author       mpv-easy
-// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/mpv-translate.js
+// @downloadURL  https://github.com/mpv-easy/mpv-easy/releases/latest/download/translate.js
 // ==/UserScript==
