+++ mpv-translate/readme.md
@@ -1,10 +1,10 @@
 ## install
 You need to install [ffmpeg](https://www.ffmpeg.org/download.html) and curl and add them to ```PATH```, curl is usually installed by default on all systems including windows11, if you only use interactive translation, you don't need to install ffmpeg
 
-- Download the latest version of [mpv-translate.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory, or use [mpsm](../mpv-mpsm/readme.md) `mpsm install mpv-translate`
+- Download the latest version of [translate.js](https://github.com/mpv-easy/mpv-easy/releases) and copy it to the mpv script directory
 - Add shortcut keys  ```ctrl+t    script-message translate``` and  ```ctrl+i    script-message interactive-translate``` to your ```input.conf```
 - If everything goes well, you can use ```ctrl+t``` to start global translation (requires ffmpeg and curl and only supports local videos), ```ctrl+i``` to start interactive translation (requires curl)
-- You can customize the styles of subtitles and tooltips, see [mpv-translate.conf](./mpv-translate.conf)
+- You can customize the styles of subtitles and tooltips, see [translate.conf](./translate.conf)
 
 ## support language
