+++ mpv-translate/translate.conf
